requirements.txt
# 手机竞品分析 - 依赖列表（Python 3.10+）
# 使用 Render/其它云 PostgreSQL 时需 psycopg2-binary（可选安装）
psycopg2-binary==2.9.9
fastapi==0.109.2
uvicorn[standard]==0.27.1
python-multipart==0.0.9
sqlalchemy==2.0.25
pandas==2.2.0
openpyxl==3.1.2
python-dotenv==1.0.1
pydantic-settings==2.2.1
httpx==0.26.0
beautifulsoup4==4.12.3
lxml==5.1.0
pdfplumber==0.11.0
pdfminer.six>=20221105
python-docx==1.1.0
